add new runtime private lib
talked with Martin and this bugger is considered private (as in: shouldn't
be used by any random kglobalaccel library consumer). It is however as ABI
stable as other frameworks, so it has symbol tracking and the whole shebang

this library also installs new headers and a cmake target. they are also
reflecting their privateness in the path/name.